requirements.txt
bcrypt==3.2.2
Flask==3.1.0
Flask_Cors==5.0.0
Flask-Executor==1.0.0
flask-marshmallow==1.3.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
lxml==5.3.0
openai==1.61.1
oss2==2.19.1
pypinyin==0.53.0
pyppeteer==2.0.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
pytz==2024.2
redis==5.2.1
Requests==2.32.3
SQLAlchemy==2.0.29
sqlalchemy_serializer==1.4.22
gunicorn==23.0.0
pymysql==1.1.1
asgiref==3.8.1
asyncio==3.4.3